Fix `shrink_to_fit` with non-empty container sizes, fixes #12
Inner map type `shrink_to_fit` requires `K: Hash + Eq` for a *reason*, it may
need to rebuild the inner map and rehash all of the values in it.

Since we provide a fake inner hasher, we must reimplement `shrink_to_fit`
ourselves in the same manner, by clearing the map and refilling the contents
using the proper hasher.